Add GUI files
this commit contains a graphical user interface for controlling the KITT
Racing car. This version of the code is not fully failproof and needs
some finishing. Manual control of the car is mostly implemented.

TODO: Add selection list from which the user can select the program it
wants to run for testing the KITT Racing car.