Update for challenge A. The way the trend is calculated is chaged a lot
KITT now stops after a turn and performs a predetermined amount of
location measurements. With some logic the faulty location measurements
are filtered out and of the remaining measurements the average is taken.
KITT then drives for approximately one fourth of the remaining distance
to the endpoint (with a minimum amount of 20 centimeter) and then again
stops to retrieve its current location.
When KITT is closer then 40 centimeters away from the endpoint this
minimum distance is decreased to ten centimeters.